(cobra bots) add a utility class to factor out the common bots features (heartbeat) and move all bots to used it + convert cobra_subscribe to be a bot and add a unittest for it

+++ test/IXCobraToStatsdBotTest.cpp
@@ -114,18 +114,18 @@ TEST_CASE("Cobra_to_statsd_bot", "[cobra_bots]")
         std::string gauge;
         std::string timer;
 
-        int sentCount = ix::cobra_to_statsd_bot(config,
-                                                channel,
-                                                filter,
-                                                position,
-                                                statsdClient,
-                                                fields,
-                                                gauge,
-                                                timer,
-                                                verbose,
-                                                maxQueueSize,
-                                                enableHeartbeat,
-                                                runtime);
+        int64_t sentCount = ix::cobra_to_statsd_bot(config,
+                                                    channel,
+                                                    filter,
+                                                    position,
+                                                    statsdClient,
+                                                    fields,
+                                                    gauge,
+                                                    timer,
+                                                    verbose,
+                                                    maxQueueSize,
+                                                    enableHeartbeat,
+                                                    runtime);
         //
         // We want at least 2 messages to be sent
         //
